Extract buildActionRegistry into src/app/actionDefinitions.ts
Move the 344-line command/action registry out of main.ts into its own
module. It takes a 19-member deps object instead of closing over the
shell; viewer and tour go through getters since both are set after the
call site. Behaviour is unchanged.

main.ts 7323 -> 6983. Baseline banked, architecture-map and
KNOWN_LIMITATIONS updated to the new count.
