+++ docs/validation/monolith-size-baseline.json
@@ -1,7 +1,7 @@
 {
   "files": {
     "src/main.ts": {
-      "lines": 7323,
+      "lines": 6983,
       "goal": 2500
     },
     "src/render/Viewer.ts": {
